Traktor: Some type conversions fixed in scene editor.

+++ code/Scene/Editor/RenderControls/CameraRenderControl.cpp
@@ -355,8 +355,8 @@ void CameraRenderControl::eventPaint(ui::PaintEvent* event)
 	}
 
 	float colorClear[4]; m_colorClear.getRGBA32F(colorClear);
-	const float deltaTime = float(m_timer.getDeltaTime());
-	const float scaledTime = m_context->getTime();
+	const double deltaTime = m_timer.getDeltaTime();
+	const double scaledTime = m_context->getTime();
 	const Matrix44 projection = m_worldRenderView.getProjection();
 	const Matrix44 view = cameraEntity->getTransform().inverse().toMatrix44();
